% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dob2months.R
\name{dob2months}
\alias{dob2months}
\title{Calculate the number of months
from date of birth to date of measurement}
\usage{
dob2months(dob, date)
}
\arguments{
\item{dob}{Date. The date of birth}

\item{date}{Date.  The date of measurement}
}
\description{
Dates should be in the format "yyyy-mm-dd".
Reformat the dates if needed.
}
\examples{
dob2months(dob= "1998-10-20", date = "2015-06-12")
}

