% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ftin2m.R
\name{ftin2m}
\alias{ftin2m}
\title{Convert feet and inches into meters
From Google convert}
\usage{
ftin2m(feet, inches = 0)
}
\arguments{
\item{feet}{Numeric. The feet part of the height}

\item{inches}{Numeric. The inch part of the height.
The default is 0}
}
\description{
Convert feet and inches into meters
From Google convert
}
\examples{
ftin2m(feet = 10, inches = 7)
}

