% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lboz2kg.R
\name{lboz2kg}
\alias{lboz2kg}
\title{Convert pounds and ounces into mkilograms
From Google convert}
\usage{
lboz2kg(pounds, ounces = 0)
}
\arguments{
\item{pounds}{Numeric. The pound part of the weight}

\item{ounces}{Numeric. The ounce part of the height.
The default is 0}
}
\description{
Convert pounds and ounces into mkilograms
From Google convert
}
\examples{
lboz2kg(pounds = 150, ounces = 5)
}

